Move theme option interface
